src/server/requirements.txt
asgiref==3.8.1
Django==4.2.13
django-cors-headers==4.3.1
djangorestframework==3.15.1
sqlparse==0.5.0
typing-extensions==4.12.1
